backend/requirements/common.txt
django~=4.0.0
djangorestframework
djangorestframework-simplejwt
django-allauth
django-cors-headers
dj-rest-auth
Pillow
black